AAPT/aapt_framework/workers/nmap_worker/requirements.txt
pika==1.3.1
neo4j==4.4.4
python-nmap==0.7.1
Flask==2.3.2 